move SharedProgramData model out of dataflow - dataflow defines own utilities for producing shared node data
